+++ Nodes/Vehicle.js
@@ -21,7 +21,7 @@ module.exports = function(Polyglot) {
 
     // polyInterface: handle to the interface
     // primary: Same as address, if the node is a primary node
-    // address: Your node address, withouth the leading 'n999_'
+    // address: Your node address, without the leading 'n999_'
     // name: Your node name
     // id is the nodedefId
     constructor(polyInterface, primary, address, name, id) {
@@ -44,34 +44,8 @@ module.exports = function(Polyglot) {
         LETSLEEP: this.onLetSleep, // call to let the vehicle sleep (disables short polling)
         HORN: this.onHorn,
         FLASH: this.onFlash,
-        LOCK: this.onLock,
-        UNLOCK: this.onUnlock,
-        SUNROOF_OPEN: this.onSunroofOpen,
-        SUNROOF_CLOSE: this.onSunroofClose,
-        PORT_OPEN: this.onPortOpen,
-        PORT_CLOSE: this.onPortClose,
         CHARGE_SET_TO: this.onChargeSetTo,
-        QUERY: this.query, // Query function from the base class
-        CLIMATE_OFF: this.onClimateOff, // stop pre-heat or pre-cool of the car
-        CLIMATE_ON: this.onClimateOn, // pre-heat or pre-cool the car
-        WINDOWS_VENT: this.onWindowsVent, // vent all the windows
-        WINDOWS_CLOSE: this.onWindowsClose, // close all the windows
-        TRUNK_OPEN: this.onTrunkOpen, // open the rear trunk
-        FRUNK_OPEN: this.onFrunkOpen, // open the front trunk (frunk)
-        HEATED_SEAT_LEVEL_DRIVER: this.onHeatedSeatDriver, // set the level on the heated seat for the driver
-        HEATED_SEAT_LEVEL_PASSENGER: this.onHeatedSeatPassenger, // set the level on the heated seat for the passenger
-        HEATED_SEAT_LEVEL_REAR_LEFT: this.onHeatedSeatRearLeft, // set the level on the heated seat for the rear left seat
-        HEATED_SEAT_LEVEL_REAR_CENTER: this.onHeatedSeatRearCenter, // set the level on the heated seat for the rear center seat
-        HEATED_SEAT_LEVEL_REAR_RIGHT: this.onHeatedSeatRearRight, // set the level on the heated seat for the rear right seat
-        SENTRY_MODE_ON: this.onSentryModeOn, // turn on Sentry Mode
-        SENTRY_MODE_OFF: this.onSentryModeOff, // turn off Sentry Mode
-        START_SOFTWARE_UPDATE: this.onStartSoftwareUpdate, // will start the car's software update if one is available.
-        MAX_DEFROST_ON: this.onMaxDefrostOn, // turns the climate control to max defrost
-        MAX_DEFROST_OFF: this.onMaxDefrostOff, // turns the climate control to the previous setting
-        CLIMATE_TEMP_SETTING_DRIVER_C: this.onSetClimateTempDriverC, // sets the climate control temp for the drivers side
-        CLIMATE_TEMP_SETTING_DRIVER_F: this.onSetClimateTempDriverF, // sets the climate control temp for the drivers side
-        CLIMATE_TEMP_SETTING_PASSENGER_C: this.onSetClimateTempPassengerC, // sets the climate control temp for the passengers side
-        CLIMATE_TEMP_SETTING_PASSENGER_F: this.onSetClimateTempPassengerF, // sets the climate control temp for the passengers side
+        QUERY_NOW: this.queryNow, // Force a query now to update the status
       };
 
       
@@ -82,8 +56,6 @@ module.exports = function(Polyglot) {
       this.drivers = {
         ST: { value: '', uom: 51 }, // SOC%
 //        GV1: { value: '', uom: 116 }, // Battery range (default mile, but UOM gathered from the vehicle)
-        GV2: { value: '', uom: 2 }, // Charge port door open
-        GV3: { value: '', uom: 2 }, // Charge port latch engaged
         GV4: { value: '', uom: 2 }, // Charge enable request
         GV5: { value: '', uom: 2 }, // Charging state
         GV6: { value: '', uom: 2 }, // Fast charger present
@@ -91,28 +63,15 @@ module.exports = function(Polyglot) {
         CC: { value: '', uom: 1 }, // Charger actual current
         CV: { value: '', uom: 72 }, // Charger voltage
         CPW: { value: '', uom: 73 }, // Charger power
-        GV8: { value: '', uom: 2 }, // Locked?
-        GV9: { value: '', uom: 51 }, // Sunroof open%
 //        GV10: { value: '', uom: 116 }, // Odometer (default mile, but multi-editor supports kilometer too)
-        GV11:  { value: '', uom: 2 }, // Sentry mode on
-//        GV12:  { value: '', uom: 4 }, // Drivers side temp
-//        GV13:  { value: '', uom: 4 }, // Passenger side temp
-//      GV14:  { value: '', uom: 4 }, // Exterior temp
-        GV15:  { value: '', uom: 2 }, // Max Defrost
-        GV17: { value: '', uom: 25 }, // Software Update Availability Status
         GV18: { value: '', uom: 2 }, // Online?
         GV19: { value: '', uom: 56 }, // Last updated unix timestamp
         GV20: { value: id, uom: 56 }, // ID used for the Tesla API
-//        CLITEMP: { value: '', uom: 4 }, // Interior temperature
-        CLIEMD: { value: '', uom: 2 }, // Climate conditioning on
-        ERR: { value: '', uom: 2 }, // In error?
+        ERR: { value: '', uom: 2 } // In error?
       };
 
       this.distance_uom = 'mi'; // defaults to miles. Pulls data from vehicle GUI to change to KM where appropriate.
-      this.temperature_uom = 'C'; // defaults to Celsius. Pulls data from vehicle GUI to change to C where appropriate.
       
-      this.drivers_temp = '15'; // need to keep these in memory for when we set one or the other
-      this.passengers_temp = '15'; // since setting one to null means the temp goes to LO
       this.let_sleep = true; // this will be used to disable short polling
     }
 
@@ -126,18 +85,6 @@ module.exports = function(Polyglot) {
         vehicleGuiSettings = await this.tesla.getVehicleGuiSettings(id);
       }
       this.vehicleUOM(vehicleGuiSettings.response);
-      logger.info('initializeUOM (%s)', this.temperature_uom);
-      if (this.temperature_uom === 'C') {
-        this.drivers.GV12 = { value: '', uom: 4 };
-        this.drivers.GV13 = { value: '', uom: 4 };
-        this.drivers.GV14 = { value: '', uom: 4 };
-        this.drivers.CLITEMP = { value: '', uom: 4 };
-      } else {
-        this.drivers.GV12 = { value: '', uom: 17 };
-        this.drivers.GV13 = { value: '', uom: 17 };
-        this.drivers.GV14 = { value: '', uom: 17 };
-        this.drivers.CLITEMP = { value: '', uom: 17 };
-      }
       
       if (this.distance_uom === 'mi') {
         this.drivers.GV1 = { value: '', uom: 116 };
@@ -200,42 +147,6 @@ module.exports = function(Polyglot) {
       await this.tesla.cmdFlashLights(id);
     }
 
-    async onLock() {
-      const id = this.vehicleId();
-      logger.info('LOCK (%s)', this.address);
-      await this.tesla.cmdDoorLock(id);
-    }
-
-    async onUnlock() {
-      const id = this.vehicleId();
-      logger.info('UNLOCK (%s)', this.address);
-      await this.tesla.cmdDoorUnlock(id);
-    }
-
-    async onSunroofOpen() {
-      const id = this.vehicleId();
-      logger.info('SUNROOF_OPEN (%s)', this.address);
-      await this.tesla.cmdSunRoof(id, 'vent');
-    }
-
-    async onSunroofClose() {
-      const id = this.vehicleId();
-      logger.info('SUNROOF_CLOSE (%s)', this.address);
-      await this.tesla.cmdSunRoof(id, 'close');
-    }
-
-    async onPortOpen() {
-      const id = this.vehicleId();
-      logger.info('PORT_OPEN (%s)', this.address);
-      await this.tesla.cmdChargePortOpen(id);
-    }
-
-    async onPortClose() {
-      const id = this.vehicleId();
-      logger.info('PORT_CLOSE (%s)', this.address);
-      await this.tesla.cmdChargePortClose(id);
-    }
-
     async onChargeSetStd() {
       const id = this.vehicleId();
       logger.info('CHARGE_SET_STD (%s)', this.address);
@@ -253,181 +164,11 @@ module.exports = function(Polyglot) {
         await this.query();
       }
 
-	async onClimateOn() {
-        const id = this.vehicleId();
-        logger.info('CLIMATE_ON (%s)', this.address);
-        await this.tesla.cmdHvacStart(id);
-        await this.query();
-      }
 
-	async onClimateOff() {
-        const id = this.vehicleId();
-        logger.info('CLIMATE_OFF (%s)', this.address);
-        await this.tesla.cmdHvacStop(id);
-        await this.query();
-      }
-
-    async onWindowsVent() {
-      const id = this.vehicleId();
-      logger.info('WINDOWS VENT (%s)', this.address);
-      await this.tesla.cmdWindows(id, 'vent');
-      await this.query();
+    async queryNow() {
+      await this.query(true);
     }
-
-    async onWindowsClose() {
-      const id = this.vehicleId();
-      logger.info('WINDOWS CLOSE (%s)', this.address);
-      await this.tesla.cmdWindows(id, 'close');
-      await this.query();
-    }
-
-    async onTrunkOpen() {
-      const id = this.vehicleId();
-      logger.info('TRUNK OPEN (%s)', this.address);
-      await this.tesla.cmdActuateTrunk(id, 'rear');
-      await this.query();
-    }
-
-    async onFrunkOpen() {
-      const id = this.vehicleId();
-      logger.info('FRUNK OPEN (%s)', this.address);
-      await this.tesla.cmdActuateTrunk(id, 'front');
-      await this.query();
-    }
-
-    async onHeatedSeatDriver(message) {
-      const id = this.vehicleId();
-
-      logger.info('SET DRIVERS HEATED SEAT (%s): %s', this.address,
-          message.value ? message.value : 'No value');
-
-      await this.tesla.cmdHeatedSeats(id, '0', message.value);
-      await this.query();
-    }
-
-    async onHeatedSeatPassenger(message) {
-      const id = this.vehicleId();
-
-      logger.info('SET PASSENGER HEATED SEAT (%s): %s', this.address,
-          message.value ? message.value : 'No value');
-
-      await this.tesla.cmdHeatedSeats(id, '1', message.value);
-      await this.query();
-    }
-
-    async onHeatedSeatRearLeft(message) {
-      const id = this.vehicleId();
-
-      logger.info('SET REAR LEFT HEATED SEAT (%s): %s', this.address,
-          message.value ? message.value : 'No value');
-
-      await this.tesla.cmdHeatedSeats(id, '2', message.value);
-      await this.query();
-    }
-
-    async onHeatedSeatRearCenter(message) {
-      const id = this.vehicleId();
-
-      logger.info('SET REAR CENTER HEATED SEAT (%s): %s', this.address,
-          message.value ? message.value : 'No value');
-
-      await this.tesla.cmdHeatedSeats(id, '4', message.value);
-      await this.query();
-    }
-
-    async onHeatedSeatRearRight(message) {
-      const id = this.vehicleId();
-
-      logger.info('SET REAR RIGHT HEATED SEAT (%s): %s', this.address,
-          message.value ? message.value : 'No value');
-
-      await this.tesla.cmdHeatedSeats(id, '5', message.value);
-      await this.query();
-    }
-
-    async onSentryModeOn() {
-      const id = this.vehicleId();
-      logger.info('SENTRY MODE ON (%s)', this.address);
-      await this.tesla.cmdSentryMode(id, 'on');
-      await this.query();
-    }
-
-    async onSentryModeOff() {
-      const id = this.vehicleId();
-      logger.info('SENTRY MODE OFF (%s)', this.address);
-      await this.tesla.cmdSentryMode(id, 'off');
-      await this.query();
-    }
-
-    async onStartSoftwareUpdate() {
-      const id = this.vehicleId();
-      logger.info('STARTING SOFTWARE UPDATE (%s)', this.address);
-      await this.tesla.cmdStartSoftwareUpdate(id);
-      await this.query();
-    }
-
-    async onMaxDefrostOn() {
-      const id = this.vehicleId();
-      logger.info('MAX DEFROST MODE ON (%s)', this.address);
-      await this.tesla.cmdMaxDefrost(id, 'on');
-      await this.query();
-    }
-
-    async onMaxDefrostOff() {
-      const id = this.vehicleId();
-      logger.info('MAX DEFROST OFF (%s)', this.address);
-      await this.tesla.cmdMaxDefrost(id, 'off');
-      await this.query();
-    }
-
-    async onSetClimateTempDriverC(message) {
-      await this.onSetClimateTempDriver(message, 'C')
-    }
-
-    async onSetClimateTempDriverF(message) {
-      await this.onSetClimateTempDriver(message, 'F')
-    }
-
-    async onSetClimateTempDriver(message, uom) {
-      const id = this.vehicleId();
-      const celsiusDeg = this.toStdTemp(message.value, uom);
-      logger.info('SETTING DRIVERS SIDE CLIMATE TEMP (%s): %s', this.address,
-        message.value ? celsiusDeg : 'No value');
-      this.drivers_temp = celsiusDeg;
-      await this.tesla.cmdSetClimateTemp(id, celsiusDeg, this.stdPassengerTemp());
-      await this.query();
-    }
-
-    // The passenger temperature is stored in GV13
-    stdPassengerTemp() {
-      const gv20 = this.getDriver('GV13'); // id used for storing the passenger temp
-      return this.toStdTemp(gv20 ? gv20.value : null, this.temperature_uom);
-    }
-
-    async onSetClimateTempPassengerC(message) {
-      await this.onSetClimateTempPassenger(message, 'C')
-    }
-
-    async onSetClimateTempPassengerF(message) {
-      await this.onSetClimateTempPassenger(message, 'F')
-    }
-
-    async onSetClimateTempPassenger(message, uom) {
-      const id = this.vehicleId();
-      const celsiusDeg = this.toStdTemp(message.value, uom);
-      logger.info('SETTING PASSENGERS SIDE CLIMATE TEMP (%s): raw %s, value %s, driver %s', this.address,
-          message.value, celsiusDeg, this.stdDriverTemp());
-      this.passengers_temp = celsiusDeg;
-      await this.tesla.cmdSetClimateTemp(id, this.stdDriverTemp(), celsiusDeg);
-      await this.query();
-    }
-
-    // The driver temperature is stored in GV12
-    stdDriverTemp() {
-      const gv12 = this.getDriver('GV12'); // id used for storing the driver temp
-      return this.toStdTemp(gv12 ? gv12.value : null, this.temperature_uom);
-    }
-
+    
     async query(longPoll) {
       const _this = this;
       if (!this.let_sleep || longPoll) {
@@ -459,57 +200,6 @@ module.exports = function(Polyglot) {
 	      logger.error('GUI Distance Units missing from gui_settings');
 	    }
 	      
-	    if (guisettings.gui_temperature_units) {
-	      this.temperature_uom = guisettings.gui_temperature_units;
-	      logger.info('Temperature Units from vehicle: %s', this.temperature_uom);
-	    } else {
-	      logger.error('GUI Temperature Units missing from gui_settings');
-	    }
-    }
-
-    celsiusToFahrenheit(celsiusDeg) {
-      return (celsiusDeg * 1.8) + 32;
-    }
-
-    fahrenheitToCelsius(fDeg) {
-      return (fDeg - 32) * 5/9;
-    }
-
-    fromStdTemp(celsiusDeg) {
-      if (this.temperature_uom === 'F') {
-        return Math.round(this.celsiusToFahrenheit(celsiusDeg)).toString();
-      } else {
-        return Math.round(celsiusDeg).toString();
-      }
-    }
-
-    toStdTemp(localDeg, uom) {
-      if (uom === 'F') {
-        return this.fahrenheitToCelsius(localDeg).toString();
-      } else {
-        return localDeg;
-      }
-    }
-
-    decodeTempUOM() {
-      if (this.temperature_uom === 'F') {
-        return 17;
-      } else {
-        return 4;
-      }
-    }
-
-    // I_SOFTWARE_UPDATE_STATUS index
-    decodSoftwareUpdateStatus(status) {
-      if (status === '') {
-        return 0
-      } else if (status === 'available') {
-        return 1
-      } else if (status === 'scheduled') {
-        return 2
-      } else if (status === 'installing') {
-        return 3
-      }
     }
 
     async queryVehicle(longPoll) {
@@ -532,7 +222,6 @@ module.exports = function(Polyglot) {
       if (vehicleData && vehicleData.response &&
         vehicleData.response.charge_state &&
         vehicleData.response.vehicle_state &&
-        vehicleData.response.climate_state &&
         vehicleData.response.gui_settings) {
 
         // logger.info('This vehicle Data %o', vehicleData);
@@ -540,19 +229,9 @@ module.exports = function(Polyglot) {
         const response = vehicleData.response;
         const chargeState = vehicleData.response.charge_state;
         const vehiculeState = vehicleData.response.vehicle_state;
-        const climateState = vehicleData.response.climate_state;
         const timestamp = Math.round((new Date().valueOf() / 1000)).toString();
 
         this.vehicleUOM(vehicleData.response.gui_settings);
-
-        if (climateState.driver_temp_setting && climateState.passenger_temp_setting) {
-          this.drivers_temp = climateState.driver_temp_setting;
-          logger.info('Drivers temp currently set to: %s', this.drivers_temp);
-          this.passengers_temp = climateState.passenger_temp_setting;
-          logger.info('Passengers temp currently set to: %s', this.passengers_temp);
-        } else {
-          logger.error('Missing drivers or passenger temp from from climate state');
-        }
 
         // We know the 'Stopped' status, but what are the others?
         if (chargeState.charging_state !== 'Stopped' &&
@@ -569,11 +248,6 @@ module.exports = function(Polyglot) {
           this.setDriver('GV1', Math.round(parseFloat(chargeState.battery_range)).toString(), true, false, 116);
         }
 
-        this.setDriver('GV2', chargeState.charge_port_door_open, false);
-        this.setDriver('GV3',
-          chargeState.charge_port_latch.toLowerCase() === 'engaged',
-          false);
-
         this.setDriver('GV4', chargeState.charge_enable_request, false);
         this.setDriver('GV5',
           chargeState.charging_state.toLowerCase() === 'charging', false);
@@ -582,10 +256,6 @@ module.exports = function(Polyglot) {
         this.setDriver('CC', chargeState.charger_actual_current, false);
         this.setDriver('CV', chargeState.charger_voltage, false);
         this.setDriver('CPW', chargeState.charger_power * 1000, false);
-        this.setDriver('GV8', vehiculeState.locked, false);
-        if (vehiculeState.sun_roof_percent_open != null) {
-        	this.setDriver('GV9', vehiculeState.sun_roof_percent_open, false);
-        }
 
         // Odometer reading
         if (this.distance_uom === 'km') {
@@ -594,39 +264,6 @@ module.exports = function(Polyglot) {
           this.setDriver('GV10', Math.round(parseFloat(vehiculeState.odometer)).toString(), true, false, 116);
         }
 
-        // Status of sentry mode.
-        if (vehiculeState.sentry_mode) {
-          this.setDriver('GV11', vehiculeState.sentry_mode, false);
-        }
-
-        // Drivers side temp
-        if (climateState.driver_temp_setting) {
-          this.setDriver('GV12', this.fromStdTemp(climateState.driver_temp_setting), true, false, this.decodeTempUOM());
-        }
-
-        // Passengers side temp
-        if (climateState.passenger_temp_setting) {
-          this.setDriver('GV13', this.fromStdTemp(climateState.passenger_temp_setting), true, false, this.decodeTempUOM());
-        }
-
-        // Exterior temp
-        if (climateState.outside_temp) {
-          this.setDriver('GV14', this.fromStdTemp(climateState.outside_temp), true, false, this.decodeTempUOM());
-        }
-
-        logger.debug("defrost_mode %s, is_front_defroster_on %s, is_auto_conditioning_on %s", climateState.defrost_mode, climateState.is_front_defroster_on, climateState.is_auto_conditioning_on);
-        // Max Defrost
-        if (climateState.defrost_mode == 2 && climateState.is_front_defroster_on && climateState.is_auto_conditioning_on) {
-          this.setDriver('GV15', true, true);
-        } else {
-          this.setDriver('GV15', false, true);
-        }
-
-        // Software Update Availability Status
-        //if (vehiculeState.software_update.status) {
-        logger.debug("software_update.status %s", vehiculeState.software_update.status);
-          this.setDriver('GV17', this.decodSoftwareUpdateStatus(vehiculeState.software_update.status), true);
-        //}
         if (this.let_sleep && !longPoll) {
           this.setDriver('GV18', false, false); // this way we know if we have to wake up the car or not
         } else {
@@ -638,11 +275,6 @@ module.exports = function(Polyglot) {
         // GV20 is not updated. This is the id we use to find this vehicle.
         // It must be already correct.
         
-        // Current temperature inside the vehicle.
-        this.setDriver('CLITEMP', this.fromStdTemp(climateState.inside_temp), true, false, this.decodeTempUOM());
-        // Status of climate conditioning.
-        this.setDriver('CLIEMD', climateState.is_climate_on, false);
-
         this.setDriver('ERR', '0', false);
         this.reportDrivers(); // Reports only changed values
       } else {